ostream inserters for various char types are deleted nowadays.
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p1423r3.html

+++ livestatus/src/Renderer.cc
@@ -65,7 +65,8 @@ void Renderer::output(const RowFragment &value) { _os << value._str; }
 
 void Renderer::output(char16_t value) {
     OStreamStateSaver s(_os);
-    _os << R"(\u)" << std::hex << std::setw(4) << std::setfill('0') << value;
+    _os << R"(\u)" << std::hex << std::setw(4) << std::setfill('0')
+        << static_cast<unsigned int>(value);
 }
 
 void Renderer::output(char32_t value) {
